Database connection done. - Add menu for added function

+++ session.model.js
@@ -1,47 +1,55 @@
 import fs from 'fs';
 
-import mongoose from 'mongoose';
 import chalk from 'chalk';
 
 import { Session } from './session.mongo.js';
 
 chalk.level = 1;
 
-// Backup all sessions from database to a timestamped file every time 
-async function backupSessions() {
+// Backup all sessions from database to a timestamped file on every startup if the file doesn't exist
+async function backupSessions(force = false) {
     const sessions = await Session.find();
     const date = new Date();
-    const dateString = `${date.getFullYear()}-${date.getMonth() + 1}-${date.getDate()}-${date.getHours()}-${date.getMinutes()}`;
-    const fileName = `./backups/sessiondata -${dateString}.json`;
-    const data = JSON.stringify(sessions, null, 4);
-    await fs
-        .promises
-        .writeFile(fileName, data, 'utf8')
-        .then(() => {
-            console.log(chalk.greenBright.italic.dim(`Successfully backed up sessions to ${fileName}`));
-        })
-        .catch((err) => {
-            console.log(chalk.redBright.italic.dim(`Error backing up sessions to ${fileName}`));
-            console.log(err);
-        });
+    const dateString = `${date.getFullYear()}-${date.getMonth() + 1}-${date.getDate()}`;
+    const fileName = `./backups/session_backup-${dateString}.json`;
+    const data = JSON.stringify(sessions);
+    if (!fs.existsSync(fileName) || force) {
+        await fs
+            .promises
+            .writeFile(fileName, data, 'utf8')
+            .then(() => {
+                console.log(chalk.greenBright.italic.dim(`Successfully backed up sessions to ${fileName}`));
+            })
+            .catch((err) => {
+                console.log(chalk.redBright.italic.dim(`Error backing up sessions to ${fileName}`));
+                console.log(err);
+            });
+    } else {
+        console.log(chalk.yellowBright.italic.dim(`Backup file ${fileName} already exists`));
+    }
 } backupSessions();
 
-// Loads the latest session from the database and returns it as an object
-async function loadSessionData(startTime) {
-    // find latest session with mongoose and return it
-    // if there are no sessions yet, create a new one
+async function initSession() {
+    let latestSession = await getSessionData();
     
     // variable with 12 hours in milliseconds
     const twelveHours = 12 * 60 * 60 * 1000;
+    if (!latestSession || Date.now() - latestSession.startTime > twelveHours) {
+        console.log(chalk.bgYellowBright.bold.underline(' Old session or no session found, creating new one '));
+        latestSession = await createNewSession();
+    } else {
+        console.log(chalk.bgGreenBright.bold.underline(' Session found, loading data '));
+        global.SESSION_START = latestSession.startTime; 
+    }
+    console.log(chalk.white.dim('Loaded session: ') + chalk.magenta.italic(latestSession.startTime));
+}
+
+
+// Loads the latest session from the database and returns it as an object
+async function getSessionData() {
     try {
-        const latestSession = await Session.findOne().sort({ startTime: -1 });        
-        if (!latestSession || Date.now() - latestSession.startTime > twelveHours) {
-            console.log(chalk.bgYellowBright.bold.underline('Old session or no session found, creating new one'));
-            const newSession = await createNewSession(startTime);
-            return newSession;
-        }
-        console.log(chalk.magenta.bold('Session restored from database for session: ') + chalk.white.underline(latestSession.startTime));
-        return latestSession;
+        let session = await Session.findOne().sort({ startTime: -1 });
+        return session;
     } catch (error) {
         console.error('Error while loading session data: ', error);
         return null;
@@ -49,12 +57,11 @@ async function loadSessionData(startTime) {
 }
 
 // 
-async function saveSessionData(sessionObj) {    
-    // upsert sessionObj as Session into database
+async function saveSessionData(session) {
     try {
         await Session.updateOne(
-            {startTime: sessionObj.startTime},
-            sessionObj,
+            {startTime: session.startTime},
+            session,
             {upsert: true}
         );
     } catch (err) {
@@ -62,18 +69,19 @@ async function saveSessionData(sessionObj) {
         console.error(err);
         return false;
     }
+    // console.log(chalk.greenBright.italic('Successfully saved session to database'));
     return true;    
 }
 
 // Creates a new session and returns it as an object
-async function createNewSession(startTime) {
+async function createNewSession(startTime = global.SESSION_START) {
     try {
         const session = new Session({
-            startTime: startTime
+            startTime: startTime // global.SESSION_START is a global variable
         });
         await session.save();
         console.log(chalk.magenta.bold('New session created for session: ') + chalk.white.underline(session.startTime));
-        cleanSessions();
+        // cleanSessions();
         return session;
     } catch (error) {
         console.error(chalk.red.underline.bold('Error while creating new session: ', error));
@@ -86,16 +94,23 @@ async function cleanSessions() {
     try {
         const sessions = await Session.find().sort({ startTime: -1 });
         if (sessions.length > 1) {
-            console.log(chalk.yellow.bold('Cleaning up garbage sessions'));
+            let counter = 0;
+            console.log(chalk.yellow.bold('Checking if any sessions can be deleted')); 
             for (let i = 1; i < sessions.length; i++) {               
                 if (sessions[i].songs.length === 0) {
                     await Session.findByIdAndDelete(sessions[i]._id);
                     console.log(chalk.red.bold.italic('Deleted session: ') + chalk.white.underline.dim.italic(sessions[i].startTime));
+                    counter++;
                 }
+            }
+            if (counter === 0) {
+                console.log(chalk.greenBright.bold('No sessions found that can be deleted'));
+            } else {
+                console.log(chalk.red.bold(`Deleted a total of ${counter} sessions`));
             }
         }
     } catch (error) {
         console.error(chalk.red.underline.bold('Error while cleaning sessions: ', error));
     }
 }
-export { loadSessionData, createNewSession, saveSessionData, };
+export { createNewSession, saveSessionData, initSession, getSessionData, backupSessions };
